+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- When doing an export with ``export_content`` and having constraints to skip
+  items, still allow to walk into subitems of the skipped ones - except for
+  skipped paths, where the whole path is skipped.
+  [thet]
 
 
 1.1 (2015-10-22)